requirements.txt
asgiref==3.9.1
Django==5.2.6
django-allauth==65.11.1
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-widget-tweaks==1.5.0
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
whitenoise==6.9.0
